+++ karchangame/src/main/java/mmud/scripting/RoomsInterface.java
@@ -31,6 +31,6 @@ public interface RoomsInterface
      * @param id the id of the room
      * @return the room or null, if not found.
      */
-    public Room find(Integer id);
+    public Room find(Long id);
 
 }
